requirements.txt
asgiref==3.3.4
backports.entry-points-selectable==1.1.0
certifi==2021.5.30
cfgv==3.3.0
charset-normalizer==2.0.4
distlib==0.3.2
Django==3.2
django-filter==2.4.0
djangorestframework==3.12.4
filelock==3.0.12
identify==2.2.12
idna==3.2
nodeenv==1.6.0
platformdirs==2.2.0
pre-commit==2.13.0
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
typing-extensions==3.7.4.3
urllib3==1.26.6
virtualenv==20.7.0
